finished the user routes
