+++ server/src/routes/api/index.ts
@@ -3,6 +3,6 @@ import { userRouter } from './user-routes.js';
 
 const router = Router();
 
-router.use('/users', userRouter);
+router.use('/user', userRouter);
 
 export default router;
